requirements.txt
Flask==3.0.0
google-generativeai
gunicorn
flask-limiter==3.5.0
bleach==6.1.0
pytest==8.1.1
flask-caching==2.1.0